+++ xharbour/contrib/hbcurl/hbcurl.c
@@ -700,7 +700,7 @@ HB_FUNC( CURL_EASY_SEND )
 
          res = curl_easy_send( hb_curl->curl, hb_parcx( 2 ), ( size_t ) hb_parclen( 2 ), &size );
 
-         hb_stornl( size, 3 );
+         hb_stornl( ( long ) size, 3 );
          /* hb_storns( size, 3 ); */
       }
 #endif
@@ -1009,7 +1009,7 @@ HB_FUNC( CURL_EASY_SETOPT )
                if( pArray )
                {
                   ULONG ulPos;
-                  ULONG ulArrayLen = hb_arrayLen( pArray );
+                  ULONG ulArrayLen = ( ULONG ) hb_arrayLen( pArray );
 
                   for( ulPos = 0; ulPos < ulArrayLen; ++ulPos )
                   {
@@ -1043,7 +1043,7 @@ HB_FUNC( CURL_EASY_SETOPT )
                if( pArray )
                {
                   ULONG ulPos;
-                  ULONG ulArrayLen = hb_arrayLen( pArray );
+                  ULONG ulArrayLen = ( ULONG ) hb_arrayLen( pArray );
 
                   for( ulPos = 0; ulPos < ulArrayLen; ++ulPos )
                      hb_curl->pHTTPHEADER = curl_slist_append( hb_curl->pHTTPHEADER, hb_arrayGetCPtr( pArray, ulPos + 1 ) );
@@ -1063,7 +1063,7 @@ HB_FUNC( CURL_EASY_SETOPT )
                if( pArray )
                {
                   ULONG ulPos;
-                  ULONG ulArrayLen = hb_arrayLen( pArray );
+                  ULONG ulArrayLen = ( ULONG ) hb_arrayLen( pArray );
 
                   for( ulPos = 0; ulPos < ulArrayLen; ++ulPos )
                      hb_curl->pHTTP200ALIASES = curl_slist_append( hb_curl->pHTTP200ALIASES, hb_arrayGetCPtr( pArray, ulPos + 1 ) );
@@ -1126,7 +1126,7 @@ HB_FUNC( CURL_EASY_SETOPT )
                if( pArray )
                {
                   ULONG ulPos;
-                  ULONG ulArrayLen = hb_arrayLen( pArray );
+                  ULONG ulArrayLen = ( ULONG ) hb_arrayLen( pArray );
 
                   for( ulPos = 0; ulPos < ulArrayLen; ++ulPos )
                      hb_curl->pMAIL_RCPT = curl_slist_append( hb_curl->pMAIL_RCPT, hb_arrayGetCPtr( pArray, ulPos + 1 ) );
@@ -1165,7 +1165,7 @@ HB_FUNC( CURL_EASY_SETOPT )
                if( pArray )
                {
                   ULONG ulPos;
-                  ULONG ulArrayLen = hb_arrayLen( pArray );
+                  ULONG ulArrayLen = ( ULONG ) hb_arrayLen( pArray );
 
                   for( ulPos = 0; ulPos < ulArrayLen; ++ulPos )
                      hb_curl->pQUOTE = curl_slist_append( hb_curl->pQUOTE, hb_arrayGetCPtr( pArray, ulPos + 1 ) );
@@ -1184,7 +1184,7 @@ HB_FUNC( CURL_EASY_SETOPT )
                if( pArray )
                {
                   ULONG ulPos;
-                  ULONG ulArrayLen = hb_arrayLen( pArray );
+                  ULONG ulArrayLen = ( ULONG ) hb_arrayLen( pArray );
 
                   for( ulPos = 0; ulPos < ulArrayLen; ++ulPos )
                      hb_curl->pPOSTQUOTE = curl_slist_append( hb_curl->pPOSTQUOTE, hb_arrayGetCPtr( pArray, ulPos + 1 ) );
@@ -1203,7 +1203,7 @@ HB_FUNC( CURL_EASY_SETOPT )
                if( pArray )
                {
                   ULONG ulPos;
-                  ULONG ulArrayLen = hb_arrayLen( pArray );
+                  ULONG ulArrayLen = ( ULONG ) hb_arrayLen( pArray );
 
                   for( ulPos = 0; ulPos < ulArrayLen; ++ulPos )
                      hb_curl->pQUOTE = curl_slist_append( hb_curl->pPREQUOTE, hb_arrayGetCPtr( pArray, ulPos + 1 ) );
@@ -1439,7 +1439,7 @@ HB_FUNC( CURL_EASY_SETOPT )
                if( pArray )
                {
                   ULONG ulPos;
-                  ULONG ulArrayLen = hb_arrayLen( pArray );
+                  ULONG ulArrayLen = ( ULONG ) hb_arrayLen( pArray );
 
                   for( ulPos = 0; ulPos < ulArrayLen; ++ulPos )
                      hb_curl->pRESOLVE = curl_slist_append( hb_curl->pRESOLVE, hb_arrayGetCPtr( pArray, ulPos + 1 ) );
@@ -1604,7 +1604,7 @@ HB_FUNC( CURL_EASY_SETOPT )
                if( pArray )
                {
                   ULONG ulPos;
-                  ULONG ulArrayLen = hb_arrayLen( pArray );
+                  ULONG ulArrayLen = ( ULONG ) hb_arrayLen( pArray );
 
                   for( ulPos = 0; ulPos < ulArrayLen; ++ulPos )
                      hb_curl->pTELNETOPTIONS = curl_slist_append( hb_curl->pTELNETOPTIONS, hb_arrayGetCPtr( pArray, ulPos + 1 ) );
