+++ xharbour/contrib/hbzebra/code93.c
@@ -117,7 +117,7 @@ static int _code93_charno( char ch )
    {
       const char * ptr = strchr( s_symbols, ch );
       if( ptr && *ptr )
-         return ptr - s_symbols + 36;
+         return (int) (ptr - s_symbols + 36);
    }
    return -1;
 }
@@ -262,7 +262,7 @@ HB_FUNC( HB_ZEBRA_CREATE_CODE93 )
    PHB_ITEM pItem = hb_param( 1, HB_IT_STRING );
    if( pItem )
    {
-      hb_zebra_ret( hb_zebra_create_code93( hb_itemGetCPtr( pItem ), hb_itemGetCLen( pItem ), hb_parni( 2 ) ) );
+      hb_zebra_ret( hb_zebra_create_code93( hb_itemGetCPtr( pItem ), (ULONG) hb_itemGetCLen( pItem ), hb_parni( 2 ) ) );
    }
    else
       hb_errRT_BASE( EG_ARG, 3012, NULL, HB_ERR_FUNCNAME, HB_ERR_ARGS_BASEPARAMS );
